# Conflicts:
#	src/main/java/com/hospital/management/service/impl/StateServiceImpl.java

+++ src/main/java/com/hospital/management/service/SystemParametersMainService.java
@@ -12,5 +12,5 @@ public interface SystemParametersMainService {
 
     List<SystemParametersMain> getSystemParametersMainList();
 
-    SystemParametersMain getSystemParametersMainListById(Integer paramsMainId);
+    SystemParametersMain getSystemParametersMainListById(Long paramsMainId);
 }
